# PERF: Optimize BitArrayExtensions.cs