--- a/techcube.uuid/Tests/techcube.uuid.PerfTests/NewInstancePerfTests.cs
+++ b/techcube.uuid/Tests/techcube.uuid.PerfTests/NewInstancePerfTests.cs
@@ -4,6 +4,7 @@ using BenchmarkDotNet.Jobs;
 
 namespace techcube.uuid.PerfTests
 {
+    [MemoryDiagnoser]
     [SimpleJob(RuntimeMoniker.Net50)]
     public class NewInstancePerfTests
     {