Hi @MrZoidberg, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
|                                Method |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Allocated |
|-------------------------------------- |-----------:|---------:|---------:|------:|--------:|-------:|----------:|
| 'UUIDv7 new instance generation test' | 3,444.1 ns | 46.73 ns | 41.42 ns | 13.92 |    0.23 | 1.1940 |   2,507 B |
|   'GUID new instance generation test' |   247.5 ns |  3.41 ns |  3.02 ns |  1.00 |    0.00 | 0.0453 |      96 B |

**Benchmarks after changes:**
|                                Method |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Allocated |
|-------------------------------------- |-----------:|---------:|---------:|------:|--------:|-------:|----------:|
| 'UUIDv7 new instance generation test' | 3,314.1 ns | 31.22 ns | 27.68 ns | 13.37 |    0.30 | 1.1139 |   2,340 B |
|   'GUID new instance generation test' |   247.8 ns |  4.92 ns |  4.60 ns |  1.00 |    0.00 | 0.0458 |      96 B |

As you can see the allocations have gone down. Please note that I made sure to replace the RandomNumberGenerator in UUIDv7.cs with a deterministic one, so that the above results are comparable. I've also made sure that all the unit tests pass following these changes as well.

Could you please confirm whether or not you think my changes are valid?

Thank you!